Update match3 examples and extensions to use actuators.
