+++ Project/Assets/ML-Agents/Examples/Match3/Prefabs/Match3Heuristic.prefab
@@ -13,8 +13,8 @@ GameObject:
   - component: {fileID: 3508723250470608012}
   - component: {fileID: 3508723250470608011}
   - component: {fileID: 3508723250470608009}
-  - component: {fileID: 3508723250470608013}
   - component: {fileID: 3508723250470608014}
+  - component: {fileID: 2112317463290853299}
   m_Layer: 0
   m_Name: Match3 Agent
   m_TagString: Untagged
@@ -51,9 +51,13 @@ MonoBehaviour:
   m_BrainParameters:
     VectorObservationSize: 0
     NumStackedVectorObservations: 1
+    m_ActionSpec:
+      m_NumContinuousActions: 0
+      BranchSizes: 
     VectorActionSize: 
     VectorActionDescriptions: []
     VectorActionSpaceType: 0
+    hasUpgradedBrainParametersWithActionSpec: 1
   m_Model: {fileID: 11400000, guid: c34da50737a3c4a50918002b20b2b927, type: 3}
   m_InferenceDevice: 0
   m_BehaviorType: 0
@@ -81,7 +85,6 @@ MonoBehaviour:
   Board: {fileID: 0}
   MoveTime: 0.25
   MaxMoves: 500
-  HeuristicQuality: 0
 --- !u!114 &3508723250470608011
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -96,7 +99,6 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   DebugMoveIndex: -1
   CubeSpacing: 1.25
-  Board: {fileID: 0}
   TilePrefab: {fileID: 4007900521885639951, guid: faee4e805953b49e688bd00b45c55f2e,
     type: 3}
 --- !u!114 &3508723250470608009
@@ -119,20 +121,6 @@ MonoBehaviour:
   BasicCellPoints: 1
   SpecialCell1Points: 2
   SpecialCell2Points: 3
---- !u!114 &3508723250470608013
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 3508723250470608007}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 08e4b0da54cb4d56bfcbae22dd49ab8d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  ActuatorName: Match3 Actuator
-  ForceHeuristic: 1
 --- !u!114 &3508723250470608014
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -147,6 +135,21 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   SensorName: Match3 Sensor
   ObservationType: 0
+--- !u!114 &2112317463290853299
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3508723250470608007}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: b17adcc6c9b241da903aa134f2dac930, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  ActuatorName: Match3 Actuator
+  ForceHeuristic: 1
+  HeuristicQuality: 0
 --- !u!1 &3508723250774301855
 GameObject:
   m_ObjectHideFlags: 0
