+++ Project/Assets/ML-Agents/Examples/Match3/Prefabs/Match3VectorObs.prefab
@@ -44,8 +44,8 @@ GameObject:
   - component: {fileID: 2118285884327540682}
   - component: {fileID: 2118285884327540685}
   - component: {fileID: 2118285884327540687}
-  - component: {fileID: 2118285884327540683}
   - component: {fileID: 2118285884327540680}
+  - component: {fileID: 3357012711826686276}
   m_Layer: 0
   m_Name: Match3 Agent
   m_TagString: Untagged
@@ -82,9 +82,13 @@ MonoBehaviour:
   m_BrainParameters:
     VectorObservationSize: 0
     NumStackedVectorObservations: 1
+    m_ActionSpec:
+      m_NumContinuousActions: 0
+      BranchSizes: 
     VectorActionSize: 
     VectorActionDescriptions: []
     VectorActionSpaceType: 0
+    hasUpgradedBrainParametersWithActionSpec: 1
   m_Model: {fileID: 11400000, guid: 9e89b8e81974148d3b7213530d00589d, type: 3}
   m_InferenceDevice: 0
   m_BehaviorType: 0
@@ -112,7 +116,6 @@ MonoBehaviour:
   Board: {fileID: 0}
   MoveTime: 0.25
   MaxMoves: 500
-  HeuristicQuality: 0
 --- !u!114 &2118285884327540685
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -127,7 +130,6 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   DebugMoveIndex: -1
   CubeSpacing: 1.25
-  Board: {fileID: 0}
   TilePrefab: {fileID: 4007900521885639951, guid: faee4e805953b49e688bd00b45c55f2e,
     type: 3}
 --- !u!114 &2118285884327540687
@@ -150,20 +152,6 @@ MonoBehaviour:
   BasicCellPoints: 1
   SpecialCell1Points: 2
   SpecialCell2Points: 3
---- !u!114 &2118285884327540683
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2118285884327540673}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 08e4b0da54cb4d56bfcbae22dd49ab8d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  ActuatorName: Match3 Actuator
-  ForceHeuristic: 0
 --- !u!114 &2118285884327540680
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -178,3 +166,18 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   SensorName: Match3 Sensor
   ObservationType: 0
+--- !u!114 &3357012711826686276
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2118285884327540673}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: b17adcc6c9b241da903aa134f2dac930, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  ActuatorName: Match3 Actuator
+  ForceHeuristic: 0
+  HeuristicQuality: 0
