+++ com.unity.ml-agents.extensions/Tests/Editor/Board/AbstractBoardTests.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using UnityEngine;
 using NUnit.Framework;
-using Unity.MLAgents.Extensions.Match3;
+using Unity.MLAgents.Extensions.Board;
 
 namespace Unity.MLAgents.Extensions.Tests.Match3
 {
