+++ com.unity.ml-agents.extensions/Tests/Editor/Board/match3obs_special0.png.meta
