+++ com.unity.ml-agents.extensions/Tests/Editor/Board/match3obs_special1.png
